pyproject.toml
[project]
name = "supply-chain-agent-demo"
version = "0.1.0"
description = "ADK multi-agent supply chain demo with A2A + MCP (BigQuery)"
requires-python = ">=3.10"
dependencies = [
  "google-adk>=0.6.0",
  "a2a-sdk>=0.3.4,<0.4.0",
  "pydantic>=2.7.0",
  "python-dotenv>=1.0.1",
  "toolbox-core>=0.5.9,<0.6.0",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["supply_chain_agent_demo*", "remote_agent*"]
